# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CAASPP & ELPAC Assessment Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /results:
    get:
      operationId: listResults
      tags: [results]
      summary: List assessment results
      parameters:
        - name: studentId
          in: query
          schema:
            type: string
        - name: assessmentId
          in: query
          schema:
            type: string
        - name: classId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Result"
    post:
      operationId: submitResult
      tags: [results]
      summary: Submit assessment result
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitResultRequest"
      responses:
        "201":
          description: Result submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Result"
  /results/{resultId}:
    get:
      operationId: getResult
      tags: [results]
      summary: Get result by ID
      parameters:
        - name: resultId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Result details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResultDetail"
  /results/{resultId}/insights:
    post:
      operationId: generateResultInsights
      tags: [results]
      summary: Generate AI insights for a specific result
      parameters:
        - name: resultId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: AI generated feedback string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResultInsightsResponse"
  /syllabus/upload:
    post:
      operationId: uploadSyllabus
      tags: [syllabus]
      summary: Upload syllabus PDF and generate questions with AI
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadSyllabusRequest"
      responses:
        "200":
          description: Generated questions from syllabus
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneratedQuestionsResponse"
...
components:
  schemas:
...
    Question:
      type: object
      properties:
        id:
          type: string
        assessmentId:
          type: string
        text:
          type: string
        audioScript:
          type: string
          nullable: true
          description: Optional transcript spoken by the AudioPlayer instead of the question text
        type:
          type: string
          enum: [multiple_choice, short_answer, essay, listening, speaking]
        options:
          type: array
          nullable: true
          items:
            type: string
        correctAnswer:
          type: string
          nullable: true
        explanation:
          type: string
          nullable: true
        points:
          type: integer
        difficulty:
          type: string
          enum: [easy, medium, hard]
        skill:
          type: string
          nullable: true
          description: ELA/Math/Science skill or ELPAC language skill
        orderIndex:
          type: integer
      required: [id, assessmentId, text, type, points, difficulty, orderIndex]
...
    Result:
      type: object
      properties:
        id:
          type: string
        assessmentId:
          type: string
        assessmentTitle:
          type: string
        assessmentType:
          type: string
          enum: [CAASPP, ELPAC]
        studentId:
          type: string
        studentName:
          type: string
        score:
          type: number
        maxScore:
          type: number
        percentage:
          type: number
        passed:
          type: boolean
        timeSpent:
          type: integer
          description: Time spent in seconds
        completedAt:
          type: string
          format: date-time
      required: [id, assessmentId, assessmentTitle, assessmentType, studentId, studentName, score, maxScore, percentage, passed, completedAt]
    ResultInsightsResponse:
      type: object
      properties:
        feedback:
          type: string
          description: Generated AI insights for the result
      required: [feedback]
    ResultDetail:
      allOf:
        - $ref: "#/components/schemas/Result"
        - type: object
          properties:
            answers:
              type: array
              items:
                $ref: "#/components/schemas/AnswerDetail"
          required: [answers]
    AnswerDetail:
      type: object
      properties:
        questionId:
          type: string
        questionText:
          type: string
        answer:
          type: string
        correctAnswer:
          type: string
          nullable: true
        isCorrect:
          type: boolean
        points:
          type: number
        maxPoints:
          type: number
      required: [questionId, questionText, answer, isCorrect, points, maxPoints]
    SubmitResultRequest:
      type: object
      properties:
        assessmentId:
          type: string
        studentId:
          type: string
        answers:
          type: array
          items:
            $ref: "#/components/schemas/AnswerSubmission"
        timeSpent:
          type: integer
      required: [assessmentId, studentId, answers, timeSpent]
    AnswerSubmission:
      type: object
      properties:
        questionId:
          type: string
        answer:
          type: string
      required: [questionId, answer]
    UploadSyllabusRequest:
      type: object
      properties:
        syllabusText:
          type: string
          description: Extracted text content from the PDF syllabus
        fileName:
          type: string
        assessmentType:
          type: string
          enum: [CAASPP, ELPAC]
        subject:
          type: string
        grade:
          type: string
        difficulty:
          type: string
          enum: [easy, medium, hard, mixed]
        questionCount:
          type: integer
          minimum: 5
          maximum: 50
        listeningRubric:
          type: string
          description: Optional custom rubric for listening questions
        readingRubric:
          type: string
          description: Optional custom rubric for reading questions
        metadata:
          type: object
          properties:
            assessmentTitle:
              type: string
            typePercentages:
              type: object
              additionalProperties:
                type: number
          description: Additional frontend-provided metadata
        writingRubric:
          type: string
          description: Optional custom rubric for writing questions
      required: [syllabusText, fileName, assessmentType, subject, grade, difficulty, questionCount]
    GeneratedQuestionsResponse:
      type: object
      properties:
        questions:
          type: array
          items:
            $ref: "#/components/schemas/Question"
        assessmentTitle:
          type: string
        summary:
          type: string
      required: [questions, assessmentTitle, summary]
...