# lib/api-zod/src/generated/types/resultInsightsResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CAASPP & ELPAC Assessment Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface ResultInsightsResponse {
  /** Generated AI insights for the result */
  feedback: string;
}