# lib/api-zod/src/generated/types/uploadSyllabusRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CAASPP & ELPAC Assessment Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { UploadSyllabusRequestAssessmentType } from "./uploadSyllabusRequestAssessmentType";
import type { UploadSyllabusRequestDifficulty } from "./uploadSyllabusRequestDifficulty";
import type { UploadSyllabusRequestMetadata } from "./uploadSyllabusRequestMetadata";

export interface UploadSyllabusRequest {
  /** Extracted text content from the PDF syllabus */
  syllabusText: string;
  fileName: string;
  assessmentType: UploadSyllabusRequestAssessmentType;
  subject: string;
  grade: string;
  difficulty: UploadSyllabusRequestDifficulty;
  /**
   * @minimum 5
   * @maximum 50
   */
  questionCount: number;
  /** Optional custom rubric for listening questions */
  listeningRubric?: string;
  /** Optional custom rubric for reading questions */
  readingRubric?: string;
  /** Additional frontend-provided metadata */
  metadata?: UploadSyllabusRequestMetadata;
  /** Optional custom rubric for writing questions */
  writingRubric?: string;
}